Open the shared color popover from a border row in the Style Library
